[Fix] fix inference api
