+++ mmedit/apis/inference.py
@@ -1,0 +1,48 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+import mmcv
+from mmcv.runner import load_checkpoint
+
+from mmedit.registry import MODELS, register_all_modules
+
+
+def delete_cfg(cfg, key='init_cfg'):
+    if key in cfg:
+        cfg.pop(key)
+    for _key in cfg.keys():
+        if isinstance(cfg[_key], mmcv.utils.config.ConfigDict):
+            delete_cfg(cfg[_key], key)
+
+
+def init_model(config, checkpoint=None, device='cuda:0'):
+    """Initialize a model from config file.
+
+    Args:
+        config (str or :obj:`mmcv.Config`): Config file path or the config
+            object.
+        checkpoint (str, optional): Checkpoint path. If left as None, the model
+            will not load any weights.
+        device (str): Which device the model will deploy. Default: 'cuda:0'.
+
+    Returns:
+        nn.Module: The constructed model.
+    """
+
+    if isinstance(config, str):
+        config = mmcv.Config.fromfile(config)
+    elif not isinstance(config, mmcv.Config):
+        raise TypeError('config must be a filename or Config object, '
+                        f'but got {type(config)}')
+    # config.test_cfg.metrics = None
+    delete_cfg(config.model, 'init_cfg')
+    print(config.model)
+    register_all_modules()
+    model = MODELS.build(config.model)
+
+    if checkpoint is not None:
+        checkpoint = load_checkpoint(model, checkpoint)
+
+    model.cfg = config  # save the config in the model for convenience
+    model.to(device)
+    model.eval()
+
+    return model
